Modernize CIK Lookup & Remove Client Base.
