+++ secedgar/filings/__init__.py
@@ -1,4 +1,3 @@
-from secedgar.filings.cik_lookup import CIKLookup  # noqa: F401
 from secedgar.filings.daily import DailyFilings  # noqa:F401
 from secedgar.filings.filing import Filing  # noqa:F401
 from secedgar.filings.filing_types import FilingType  # noqa:F401
